chore: move v2-to-v3 getters to utils/get
