+++ src/transforms/v2-to-v3/utils/addV3ClientRequire.ts
@@ -2,8 +2,7 @@ import { Collection, Identifier, JSCodeshift, ObjectPattern, Property } from "js
 
 import { AddV3ClientModuleOptions } from "./addV3ClientModule";
 import { PACKAGE_NAME } from "./config";
-import { getRequireVariableDeclaration } from "./getRequireVariableDeclaration";
-import { getV2ServiceModulePath } from "./getV2ServiceModulePath";
+import { getRequireVariableDeclaration, getV2ServiceModulePath } from "./get";
 
 export const addV3ClientRequire = (
   j: JSCodeshift,
