+++ src/transforms/v2-to-v3/utils/get/getV2ClientIdentifiers.ts
