+++ src/transforms/v2-to-v3/utils/get/getV2ServiceModulePath.ts
@@ -1,4 +1,4 @@
-import { PACKAGE_NAME } from "./config";
+import { PACKAGE_NAME } from "../config";
 
 export const getV2ServiceModulePath = (v2ClientName: string) =>
   `${PACKAGE_NAME}/clients/${v2ClientName.toLowerCase()}`;
