+++ src/transforms/v2-to-v3/utils/get/getV3ClientName.ts
@@ -1,4 +1,4 @@
-import { CLIENT_NAMES_MAP } from "./config";
+import { CLIENT_NAMES_MAP } from "../config";
 
 // Returns v3 client name for the provided v2 client name.
 export const getV3ClientName = (clientName: string) => {
