+++ src/transforms/v2-to-v3/utils/get/getV3ClientPackageName.spec.ts
@@ -1,23 +1,26 @@
-import { CLIENT_NAMES_MAP } from "./config";
-import { getV3ClientName } from "./getV3ClientName";
+import { CLIENT_PACKAGE_NAMES_MAP } from "../config";
+import { getV3ClientPackageName } from "./getV3ClientPackageName";
 
-describe(getV3ClientName.name, () => {
-  it.each(Object.entries(CLIENT_NAMES_MAP))("getV3ClientName('%s') === '%s'", (input, output) => {
-    expect(getV3ClientName(input)).toBe(output);
-  });
+describe(getV3ClientPackageName.name, () => {
+  it.each(Object.entries(CLIENT_PACKAGE_NAMES_MAP))(
+    "getClientName('%s') === '%s'",
+    (input, output) => {
+      expect(getV3ClientPackageName(input)).toBe(`@aws-sdk/${output}`);
+    }
+  );
 
   it.each(["ImportExport", "MobileAnalytics", "SimpleDB"])(
     "throws for deprecated client '%s'",
     (deprecatedClient) => {
       expect(() => {
-        getV3ClientName(deprecatedClient);
+        getV3ClientPackageName(deprecatedClient);
       }).toThrow(new Error(`Client '${deprecatedClient}' is either deprecated or newly added.`));
     }
   );
 
   it.each(["UNDEFINED", "NULL", "UNKNOWN"])("throws for unknown client '%s'", (unknownClient) => {
     expect(() => {
-      getV3ClientName(unknownClient);
+      getV3ClientPackageName(unknownClient);
     }).toThrow(new Error(`Client '${unknownClient}' is either deprecated or newly added.`));
   });
 });
