+++ src/transforms/v2-to-v3/utils/index.ts
@@ -1,9 +1,5 @@
 export * from "./addV3ClientModule";
 export * from "./containsRequire";
-export * from "./getClientMetadata";
-export * from "./getV2ClientNames";
-export * from "./getV2DefaultModuleName";
-export * from "./getV2ServiceModuleNames";
 export * from "./isTypeScriptFile";
 export * from "./removeDefaultModuleIfNotUsed";
 export * from "./removePromiseCalls";
